% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{ChooseAlgorithm}
\alias{ChooseAlgorithm}
\title{Returns which algorithm among Multiclass SVM, KNN and Multiclass logistic  performs better on the labelled data set using cross validation technique}
\usage{
ChooseAlgorithm(Xtilde, Y, K = NULL)
}
\arguments{
\item{Xtilde}{n * p centered and scaled data points}

\item{Y}{A numeric n vector of labels of Xtilde}

\item{K}{Number of classes, ( If K is NULL, it calculates K as maximum entry in Y)}
}
\value{
A string among c("KNN", "Logistic", "SVM") which tells us which algorithm performs better on the labelled data set
}
\description{
Returns which algorithm among Multiclass SVM, KNN and Multiclass logistic  performs better on the labelled data set using cross validation technique
}
\examples{
X <- rbind(matrix(rnorm(10,0,1),5,2),matrix(rnorm(10,1,2),5,2), matrix(rnorm(10,4,3),5,2))
fold_ids <- sample(1:15)
X <- X[fold_ids, ]
Xtilde <- standardizeX(X)$tilde
Y <- c(1,1,1,1,1,2,2,2,2,2,3,3,3,3,3)
Y <- Y[fold_ids]
out <- ChooseAlgorithm(Xtilde,Y,3)
#"KNN"
}
