% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{Comparison_error}
\alias{Comparison_error}
\title{Comparison error compares the error of unlabeled data points before labeling L points and after labeling L points that were given by SSLBudget function}
\usage{
Comparison_error(X, Y, Z, W, new_data, K = NULL)
}
\arguments{
\item{X}{n1 * p matrix of labelled data points. Here the rows contains the n1 labelled data points. Each data point belongs to R^p}

\item{Y}{n1 vector of labels for data points in X}

\item{Z}{m1*p matrix of Unlabeled data points}

\item{W}{m1 vector of true labels of data points in Z}

\item{new_data}{L*(p+1) matrix of data points that were given by SSLBudget appended with their true labels}

\item{K}{number of classes}
}
\value{
A list with two components, error_before - hamming distance between true labels and predicted labels before user labels L data points suggested by SSLBudget, error_after - hamming distance between true labels and predicted labels after user labels L data points suggested by SSLBudget
}
\description{
Comparison error compares the error of unlabeled data points before labeling L points and after labeling L points that were given by SSLBudget function
}
\examples{

}
