% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{SSLBudget}
\alias{SSLBudget}
\title{SSLBudget outputs what unlabeled data points should be labelled based on the budget}
\usage{
SSLBudget(X, Y, L, Z, K = NULL)
}
\arguments{
\item{X}{n1 * p matrix of labelled data points. Here the rows contains the n1 labelled data points. Each data point belongs to R^p}

\item{Y}{n1 vector of labels for data points in X}

\item{L}{positive integer - number of data points that user can label}

\item{Z}{m1*p matrix of Unlabeled data points}

\item{K}{number of classes}
}
\value{
Tolabel - A L*p matrix of data points from Z that user should label
}
\description{
SSLBudget outputs what unlabeled data points should be labelled based on the budget
}
