% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{SSLconf}
\alias{SSLconf}
\title{SSLconf function arranges the unlabeled data points based on the confidence of prediction}
\usage{
SSLconf(X, Y, Z, K = NULL)
}
\arguments{
\item{X}{n1 * p matrix of labelled data points. Here the rows contains the n1 labelled data points. Each data point belongs to R^p}

\item{Y}{n1 vector of labels for data points in X}

\item{Z}{m1*p matrix of Unlabeled data points}

\item{K}{number of classes}
}
\value{
A list with five components, conf - A matrix with data points of Z arranged from high to low confidence of prediction, High - A matrix containing data points of Z that are predicted with high confidence, Low - A matrix containing data points of Z that are predicted with low confidence, Average - A matrix containing data points of Z that are predicted with moderate confidence, Remat - A matrix containing data points of Z that should possibly be relabeled
}
\description{
SSLconf function arranges the unlabeled data points based on the confidence of prediction
}
\examples{
data = iris
data = data[sample(1:150), ]
X = data[1:25,1:4]
Y = data[1:25,5]
Z = data[26:150, 1:4]
K = 3
SSLconf(X, Y, Z, K)

}
