% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{SSLpredict}
\alias{SSLpredict}
\title{SSLpredict predicts the labels for unlabeled data}
\usage{
SSLpredict(X, Y, Z, K = NULL)
}
\arguments{
\item{X}{n * p matrix of labelled data points. Here the rows contains the n labelled data points. Each data point belongs to R^p}

\item{Y}{A numeric n vector of labels of Xtilde}

\item{Z}{m * p matrix of Unlabeled data points}

\item{K}{Number of classes, ( If K is NULL, it calculates K as maximum entry in Y)}
}
\value{
A m1 vector of predicted labels for data points in Z
}
\description{
SSLpredict predicts the labels for unlabeled data
}
\examples{
 X <- rbind(matrix(rnorm(10,0,1),5,2),matrix(rnorm(10,1,2),5,2), matrix(rnorm(10,4,3),5,2))
fold_ids <- sample(1:15)
X <- X[fold_ids, ]
Y <- c(1,1,1,1,1,2,2,2,2,2,3,3,3,3,3)
Y <- Y[fold_ids]
Z <- rbind(matrix(rnorm(80,0,1),40,2),matrix(rnorm(80,1,2),40,2), matrix(rnorm(80,4,3),40,2))
Z <- Z[sample(1:120), ]
out <- SSLpredict(X, Y, Z,3)
}
