% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{standardizeX}
\alias{standardizeX}
\title{standardizeX function centers and scales data}
\usage{
standardizeX(X)
}
\arguments{
\item{X}{A matrix to center and scale}
}
\value{
A list of 3 elements, $tilde contains the center and scaled matrix, $weights returns  sqrt(X_j^{\top}X_j/n) after centering of X but before scaling and $means returns means of columns of X
}
\description{
standardizeX function centers and scales data
}
\examples{
X <- matrix(rnorm(20,2,1),5,4)
standardizeX(X)
#$ tilde
#[,1]       [,2]       [,3]       [,4]
#[1,]  1.6882137 -0.7593179 -0.6990462  1.8321703
#[2,]  0.1689194  0.2360252 -0.1677231 -0.4757201
#[3,] -0.8734264  0.1447277 -0.8047698 -1.1429930
#[4,]  0.1694665 -1.2727583 -0.2684332  0.1025888
#[5,] -1.1531731  1.6513233  1.9399723 -0.3160460
#$weights
#[1] 0.9099549 0.8239372 0.6899589 0.8253017
#$means
#[1] 2.446213 1.150378 2.978231 1.841284
}
