% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLogistic_function.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Title}
\usage{
LRMultiClass(
  X,
  y,
  Xt,
  yt,
  numIter = 50,
  eta = 0.1,
  lambda = 1,
  beta_init = NULL
)
}
\arguments{
\item{beta_init}{}
}
\value{

}
\description{
Title
}
