% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyKmeansfunction.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Function that implements K-means algorithm.}
\usage{
MyKmeans(X, K)
}
\arguments{
\item{X}{n * p matrix of data points}

\item{K}{Number of classes}
}
\value{
n vector of cluster labels
}
\description{
Function that implements K-means algorithm.
}
\examples{
X <- matrix(c(12, 5, 2, 3, 1, 3, 4, 4, 2, 5, 5, 7), 4, 3)
MyKmeans(X, K = 2)
# output:
#      [,1] [,2] [,3] [,4]
# [1,]    1    2    2    2
}
