% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{SSLconf}
\alias{SSLconf}
\title{SSLconf divides the unlabeled data into matrices "High", "Low", "Average" based on the confidence in predicting them.}
\usage{
SSLconf(X, Y, Z, K = NULL, Plots = T)
}
\arguments{
\item{X}{n * p matrix of labelled data points. Here the rows contains the n labelled data points. Each data point belongs to R^p}

\item{Y}{A numeric n vector of labels of Xtilde}

\item{Z}{m * p matrix of Unlabeled data points}

\item{K}{Number of classes, ( If K is NULL, it calculates K as maximum entry in Y)}

\item{Plots}{Takes argument TRUE or FALSE, depending on whether  the function should output plots. Default is TRUE}
}
\value{
A list with four elements,
\item{High}{A matrix containing data points of Z that are predicted with high confidence}
\item{Low}{A matrix containing data points of Z that are predicted with low confidence}
\item{Average}{A matrix containing data points of Z that are predicted with moderate confidence}
\item{Remat}{A matrix containing data points of Z that should possibly be relabeled}
}
\description{
SSLconf divides the unlabeled data into matrices "High", "Low", "Average" based on the confidence in predicting them.
}
\examples{
X <- rbind(matrix(rnorm(10,0,1),5,2),matrix(rnorm(10,1,2),5,2), matrix(rnorm(10,4,3),5,2))
fold_ids <- sample(1:15)
X <- X[fold_ids, ]
Y <- c(1,1,1,1,1,2,2,2,2,2,3,3,3,3,3)
Y <- Y[fold_ids]
Z <- rbind(matrix(rnorm(80,0,1),40,2),matrix(rnorm(80,1,2),40,2), matrix(rnorm(80,4,3),40,2))
Z <- Z[sample(1:120), ]
out <- SSLconf(X, Y, Z,3)


}
