% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{SSLconf}
\alias{SSLconf}
\title{SSLconf divides the unlabeled data into matrices "High", "Low", "Average" based on the confidence in predicting them.}
\usage{
SSLconf(X, Y, Z, K = NULL)
}
\arguments{
\item{X}{n * p matrix of labelled data points. Here the rows contains the n labelled data points. Each data point belongs to R^p}

\item{Y}{numeric n vector of labels for data points in X}

\item{Z}{m * p matrix of Unlabeled data points}

\item{K}{number of classes(If K is null, It calculates K as the maximum entry in Y)}
}
\value{
A list with four elements, $High - A matrix containing data points of Z that are predicted with high confidence, $Low - A matrix containing data points of Z that are predicted with low confidence, $Average - A matrix containing data points of Z that are predicted with moderate confidence, $Remat - A matrix containing data points of Z that should possibly be relabeled
}
\description{
SSLconf divides the unlabeled data into matrices "High", "Low", "Average" based on the confidence in predicting them.
}
\examples{
X <- rbind(matrix(rnorm(10,0,1),5,2),matrix(rnorm(10,1,2),5,2), matrix(rnorm(10,4,3),5,2))
fold_ids <- sample(1:15)
X <- X[fold_ids, ]
Y <- c(1,1,1,1,1,2,2,2,2,2,3,3,3,3,3)
Y <- Y[fold_ids]
Z <- rbind(matrix(rnorm(80,0,1),40,2),matrix(rnorm(80,1,2),40,2), matrix(rnorm(80,4,3),40,2))
Z <- Z[sample(1:120), ]
out <- SSLconf(X, Y, Z,3)


}
