% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm.R
\name{SSLpredict}
\alias{SSLpredict}
\title{SSLpredict predicts the labels for unlabeled data}
\usage{
SSLpredict(X, Y, Z, K = NULL)
}
\arguments{
\item{X}{n1 * p matrix of labelled data points. Here the rows contains the n1 labelled data points. Each data point belongs to R^p}

\item{Y}{n1 vector of labels for data points in X}

\item{Z}{m1*p matrix of Unlabeled data points}

\item{K}{number of classes}
}
\value{
A m1 vector of predicted labels for data points in Z
}
\description{
SSLpredict predicts the labels for unlabeled data
}
\examples{
#'
data(iris)
data1 = iris[sample(1:150), ]
X = data1[1:25,1:4]
Y = data1[1:25,5]
Z = data1[26:150, 1:4]
SSLpredict(X, Y,  Z) # 1
}
